requirements.txt
streamlit
pandas
yfinance
alpaca-py